requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-cloud-aiplatform==1.67.1
google-cloud-firestore==2.18.0
vertexai==1.67.1
python-dotenv==1.0.1
pydantic==2.9.2